grpc + product count from int16 to uint32

+++ internal/domain/model/product.go
@@ -4,5 +4,5 @@ type Product struct {
 	Sku   uint64
 	Price float64
 	Name  string
-	Count int16
+	Count uint32
 }
